zongsan/yolov5-master/requirements--安装文件.txt
albumentations==1.4.11
comet_ml==3.43.2
coremltools==7.2
Flask==3.0.3
GitPython==3.1.43
GitPython==3.1.43
ipython==8.12.3
matplotlib==3.7.1
mss==9.0.1
nncf==2.11.0
numpy==2.0.0
onnx==1.16.1
onnxruntime==1.18.1
onnxsim==0.4.36
opencv_contrib_python==4.9.0.80
opencv_python==4.8.0.74
openvino==2024.2.0
paddlepaddle_gpu==2.5.2.post102
pafy==0.5.5
pandas==2.2.2
Pillow==10.4.0
psutil==5.9.5
pycocotools==2.0.8
PyQt5==5.15.10
PyQt5_sip==12.12.1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
scipy==1.14.0
seaborn==0.13.2
setuptools==67.8.0
tensorflow==2.14.0
tensorflow_intel==2.14.0
tensorflowjs==4.20.0
tensorrt==10.2.0
tflite_runtime==2.14.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.2
tritonclient==2.47.0
ultralytics==8.2.48
ultralytics_thop==2.0.0
x2paddle==1.4.1
